Perf Utf8Encoding for ARM64 comparison
