+++ src/benchmarks/micro/libraries/Common/Chinese.txt
